refactor: create test_hydro with kge, nse and pbias tests
